Rename FileStorage and FileHandler classes
