+++ src/hardware/components/drive/AbstractDrive.java
@@ -1,17 +1,17 @@
 package hardware.components.drive;
 
-import hardware.components.shared.FileHandler;
 import hardware.components.shared.FileStorage;
+import hardware.components.shared.FileHandler;
 import hardware.components.shared.enums.ComponentType;
 import hardware.components.shared.enums.StorageCapacity;
 import software.file.File;
 
-public abstract class AbstractDrive implements Drive, FileHandler {
-    private final FileStorage fileStorage;
+public abstract class AbstractDrive implements Drive, FileStorage {
+    private final FileHandler fileHandler;
     private final String name;
 
     public AbstractDrive(StorageCapacity storageCapacity, String name) {
-        this.fileStorage = new FileStorage(storageCapacity);
+        this.fileHandler = new FileHandler(storageCapacity);
         this.name = name;
     }
     //  TODO: metoda do zaimplementowania w kolejnym tasku
@@ -27,22 +27,22 @@ public abstract class AbstractDrive implements Drive, FileHandler {
 
     @Override
     public void addFile(File file) {
-        fileStorage.addFile(file);
+        fileHandler.addFile(file);
     }
 
     @Override
     public void listFiles() {
-        fileStorage.listFiles();
+        fileHandler.listFiles();
     }
 
     @Override
     public void removeFile(File file) {
-        fileStorage.removeFile(file);
+        fileHandler.removeFile(file);
     }
 
     @Override
     public File findFile(String fileName) {
-        return fileStorage.findFile(fileName);
+        return fileHandler.findFile(fileName);
     }
 
     @Override
