+++ src/hardware/components/shared/FileStorage.java
@@ -1,50 +1,10 @@
 package hardware.components.shared;
 
-import hardware.components.shared.enums.StorageCapacity;
 import software.file.File;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.NoSuchElementException;
-
-public class FileStorage {
-
-    private final StorageCapacity storageCapacity;
-    private long currentCapacity = 0;
-    private List<File> files = new ArrayList<>();
-
-    public FileStorage(StorageCapacity storageCapacity) {
-        this.storageCapacity = storageCapacity;
-    }
-
-    public long getEmptyCapacity() {
-        return storageCapacity.getSize() - currentCapacity;
-    }
-
-    public void addFile(File file) {
-        files.add(file);
-        currentCapacity += file.getSize();
-    }
-
-    public void removeFile(File file) {
-        if (files.contains(file)) {
-            files.remove(file);
-            currentCapacity -= file.getSize();
-        }
-    }
-
-    public File findFile(String fileName) {
-        return files.stream()
-                .filter(f -> f.getName().equals(fileName))
-                .findFirst()
-                .orElseThrow(() -> new NoSuchElementException("File not found"));
-    }
-
-    public void listFiles() {
-//         files.forEach(System.out::println);
-        for (File file : files) {
-            System.out.println(file.getName());
-        }
-    }
-
+public interface FileStorage {
+    void addFile(File file);
+    void removeFile(File file);
+    void listFiles();
+    File findFile(String name);
 }
